Rename and move to domain layer
