+++ source/Processing.IntegrationTests/Application/MoveIn/MoveInTests.cs
@@ -19,6 +19,7 @@ using Processing.Application.ChangeOfSupplier.Validation;
 using Processing.Application.Common;
 using Processing.Application.MoveIn;
 using Processing.Domain.Consumers;
+using Processing.Domain.EnergySuppliers.Errors;
 using Processing.Domain.SeedWork;
 using Xunit;
 using Xunit.Categories;
@@ -43,7 +44,7 @@ namespace Processing.IntegrationTests.Application.MoveIn
 
             var result = await SendRequestAsync(request).ConfigureAwait(false);
 
-            AssertValidationError<EnergySupplierMustBeKnownRuleError>(result);
+            AssertValidationError<UnknownEnergySupplier>(result);
             Assert.False(result.Success);
         }
